+++ src/App.jsx
@@ -1,6 +1,9 @@
 import { Route, BrowserRouter as Router, Routes } from 'react-router-dom'
 
 import Home from '@/pages/Home'
+import CareersHub from '@/pages/CareersHub'
+import ConnectionsPage from '@/pages/Connections'
+import MediaPage from '@/pages/Media'
 import PreviousEvents from '@/pages/PreviousEvents'
 import PreviousEvent from '@/pages/PreviousEvent'
 import NotFound from '@/pages/NotFound'
@@ -18,6 +21,9 @@ function App() {
         </a>
         <Routes>
           <Route path="/" element={<Home />} />
+          <Route path="/careers-hub" element={<CareersHub />} />
+          <Route path="/connections" element={<ConnectionsPage />} />
+          <Route path="/media" element={<MediaPage />} />
           <Route path="/previous-events" element={<PreviousEvents />} />
           <Route path="/previous-events/:year" element={<PreviousEvent />} />
           <Route path="*" element={<NotFound />} />