+++ src/pages/Home.jsx
@@ -1,27 +1,29 @@
 import { PageLayout } from '@/layouts/PageLayout'
 import TeamSection from '@/layouts/TeamSection'
-import JobBoardSection from '@/layouts/JobBoardSection'
+import CommunitySection from '@/layouts/CommunitySection'
 import LandingSection from '@/layouts/LandingSection'
 import LocationSection from '@/layouts/LocationSection'
+import MembersSection from '@/layouts/MembersSection'
 import PartnersSection from '@/layouts/PartnersSection'
 import SessionsSection from '@/layouts/SessionsSection'
 import SpeakersSection from '@/layouts/SpeakersSection'
 
-import { partnersData } from '@/data/2025/partners'
-import { SpeakersData as Speakers2025 } from '@/data/2025/speakers'
-import { teamData } from '@/data/2025/team'
+import { partnersData } from '@/data/2026/partners'
+import { SpeakersData as Speakers2026 } from '@/data/2026/speakers'
+import { teamData } from '@/data/2026/team'
 
 function Home() {
+  const currentYear = new Date().getFullYear()
   return (
     <PageLayout>
       <h1 id="main-heading" className="sr-only">
-        Michigan DevFest 2025
+        Compass Detroit {currentYear}
       </h1>
       <LandingSection />
       <LocationSection />
       <SessionsSection
-        year={2025}
-        speakersData={Speakers2025}
+        year={currentYear}
+        speakersData={Speakers2026}
         tracks={[
           'Schedule',
           'Map',
@@ -33,16 +35,14 @@ function Home() {
           'Innovation',
           'Workshops',
         ]}
-        defaultExpanded
       />
-      <SpeakersSection
-        year={2025}
-        speakersData={Speakers2025}
-        defaultExpanded
-      />
-      <JobBoardSection year={2025} />
-      <PartnersSection year={2025} partnersData={partnersData} />
-      <TeamSection year={2025} teamData={teamData} />
+      <CommunitySection year={currentYear} />
+      <MembersSection year={currentYear} />
+
+      <SpeakersSection year={currentYear} speakersData={Speakers2026} />
+
+      <PartnersSection year={currentYear} partnersData={partnersData} />
+      <TeamSection year={currentYear} teamData={teamData} />
     </PageLayout>
   )
 }